# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Persian CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, unpaid]
        - name: customer_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/pay:
    patch:
      operationId: payInvoice
      tags: [invoices]
      summary: Mark invoice as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/copy:
    post:
      operationId: copyInvoice
      tags: [invoices]
      summary: Copy an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: New copied invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List all suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a new supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
        "404":
          description: Not found
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /panels:
    get:
      operationId: listPanels
      tags: [panels]
      summary: List all panels
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of panels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Panel"
    post:
      operationId: createPanel
      tags: [panels]
      summary: Create a panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelInput"
      responses:
        "201":
          description: Created panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
  /panels/{id}:
    get:
      operationId: getPanel
      tags: [panels]
      summary: Get panel by ID with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Panel with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanelWithItems"
        "404":
          description: Not found
    patch:
      operationId: updatePanel
      tags: [panels]
      summary: Update a panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelInput"
      responses:
        "200":
          description: Updated panel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
    delete:
      operationId: deletePanel
      tags: [panels]
      summary: Delete a panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /panels/{id}/items:
    post:
      operationId: addPanelItem
      tags: [panels]
      summary: Add item to panel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelItemInput"
      responses:
        "201":
          description: Created panel item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanelItem"
  /panels/{id}/items/{itemId}:
    patch:
      operationId: updatePanelItem
      tags: [panels]
      summary: Update panel item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PanelItemInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PanelItem"
    delete:
      operationId: deletePanelItem
      tags: [panels]
      summary: Delete panel item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    InvoiceItem:
      type: object
      required: [title, quantity, unit_price]
      properties:
        title:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        brand:
          type: ["string", "null"]
        equipment_id:
          type: ["integer", "null"]
        supplier_id:
          type: ["integer", "null"]
        purchase_price:
          type: ["number", "null"]
    Invoice:
      type: object
      required: [id, customer_id, title, date, items, total_amount, status, created_at]
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        customer_name:
          type: ["string", "null"]
        job_id:
          type: ["integer", "null"]
        title:
          type: string
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        notes:
          type: ["string", "null"]
        discount:
          type: number
        total_amount:
          type: number
        status:
          type: string
          enum: [paid, unpaid]
        created_at:
          type: string
    InvoiceInput:
      type: object
      required: [customer_id, title, date, items]
      properties:
        customer_id:
          type: integer
        job_id:
          type: ["integer", "null"]
        title:
          type: string
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        notes:
          type: string
        discount:
          type: number
        status:
          type: string
          enum: [paid, unpaid]
    InvoiceUpdate:
      type: object
      properties:
        customer_id:
          type: integer
        job_id:
          type: ["integer", "null"]
        title:
          type: string
        date:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        notes:
          type: ["string", "null"]
        discount:
          type: number
        status:
          type: string
          enum: [paid, unpaid]
...
    Supplier:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        contact_person:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        created_at:
          type: string
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contact_person:
          type: string
        phone:
          type: string
        address:
          type: string
        description:
          type: string
...
    Panel:
      type: object
      required: [id, name, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        created_at:
          type: string
    PanelItem:
      type: object
      required: [id, panel_id, equipment_id, quantity]
      properties:
        id:
          type: integer
        panel_id:
          type: integer
        equipment_id:
          type: integer
        equipment_name:
          type: ["string", "null"]
        quantity:
          type: number
    PanelWithItems:
      allOf:
        - $ref: "#/components/schemas/Panel"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/PanelItem"
    PanelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    PanelItemInput:
      type: object
      required: [equipment_id, quantity]
      properties:
        equipment_id:
          type: integer
        quantity:
          type: number