# lib/api-zod/src/generated/types/invoiceUpdate.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Persian CRM API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceItem } from './invoiceItem.ts';
import type { InvoiceUpdateStatus } from './invoiceUpdateStatus.ts';

export interface InvoiceUpdate {
  customer_id?: number;
  /** @nullable */
  job_id?: number | null;
  title?: string;
  date?: string;
  items?: InvoiceItem[];
  /** @nullable */
  notes?: string | null;
  discount?: number;
  status?: InvoiceUpdateStatus;
}