# lib/api-zod/src/generated/types/panelInput.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Persian CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PanelInput {
  name: string;
  description?: string;
}