# lib/api-zod/src/generated/types/panelItem.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * Persian CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PanelItem {
  id: number;
  panel_id: number;
  equipment_id: number;
  /** @nullable */
  equipment_name?: string | null;
  quantity: number;
}